Changed bash shortcuts and added backlight ones.
cl now accepts trailing arguments which are forwarded to ls, allowing
e.g. `cl someDirectory -a`, which shows all files on entering the
directory. mkcd changed to mkcl. back changed to a function and all
arguments are forwarded to the ls call in cl. Added mvcl to move a
file and follow it if moved to a directory.

Added bash shortcuts for backlight control to avoid having to search
for F keys in the dark.